pyproject.toml
[project]
name = "mit-6-824"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=40.11.0",
    "faker-commerce>=1.0.4",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
]
